+++ app/src/app.config.ts
@@ -1,6 +1,8 @@
 export default defineAppConfig({
   pages: [
-    'pages/index/index'
+    'pages/index/index',
+    'pages/items/index',
+    'pages/item-edit/index'
   ],
   window: {
     backgroundTextStyle: 'light',